Adds rate limiting for API routes
Installed express-rate-limit. This library will allow us to set the rate
limiting for our server. It is currently configured
to allow up to 100 requests per 15 min from a client. This rate limit
will only be applied to routes that contain `/api`. The setting `trust
proxy` is enabled on line 14 because it's the recommended setting when server is hosted Heroku.