Fix 500 errors and return expected_delivery_date
500 errors initally encountered during testing was due to bad test
input. This commit should make a few fixes to make tests valid. I had
to make /createAddress return the address_id and store it in `from` var so var could be used in
/createLetter to pass. While not an ideal solution, it's not terrible.
We just need to be careful when using `from` after that test.

/createLetter now returns expected_delivery_date so the frontend can
show users this after successful redirect.