Issue refund if LobAPI throws error /createLetter
This commit adds the ability to refund users if their letter doesn't
send. Right now, there's no way to test it. The next step is to utilize
the stripe API test env in these docs https://stripe.com/docs/testing

Right now, it's untested but should work according to docs.

Eventually we need to add error handling for refund. Right now, not
likely to be used, but we should have it as the project scales.