Address verification and test for /createAddress
Adds address verification for /createAddress route. I removed the
empty string validation found in /addressVerification route as the lob
api checks for it and I couldn't seem to get it to play nice with rest
of code. Only had time to write single test for it, but more tests seem
unneccesary as the logic has already been tested for
/addressVerification.